+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Maintenance
 * Fixed multiple forbidden api warnings from the code []()
+* Re-enable `SegmentReplicationIT` for the remote index build test suite and replace the fixed sleep with a replication wait [#3550](https://github.com/opensearch-project/k-NN/pull/3550)
 
 ### Bug Fixes
 * Fix native thread leak in Lucene HNSW merge executor when index_thread_qty > 1 [#3102](https://github.com/opensearch-project/k-NN/issues/3102)